requirements_dev.txt
-r requirements.txt
pytest>=7.0.0
pytest-homeassistant-custom-component>=0.13.0
flake8>=6.0.0
black>=23.1.0
isort>=5.12.0
pre-commit>=3.1.0
pylint>=2.16.0
pytouchline_extended==0.4.5 